# swagger.yaml
openapi: 3.0.3
info:
  title: Admin Authentication API
  description: API for admin user authentication, including signup, login, password reset, logout functionality, and comment management.
  version: 1.0.0
  contact:
    name: Developer Support
    email: support@example.com
    url: https://example.com/support
servers:
  - url: http://localhost:2670
    description: Local development server
  - url: https://my-blog-2vl8.onrender.com
    description: Online Render server
paths:
...
  /blogPost:
    post:
      summary: Create a new blog post (admin only)
      description: Allows an admin user to create a new blog post with an optional image or video upload.
      tags:
        - Blog Posts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the blog post.
                  example: "Latest Football Updates"
                description:
                  type: string
                  description: The description of the blog post.
                  example: "This is an update about recent football matches."
                category:
                  type: string
                  description: The category of the blog post.
                  example: football
                  enum: [npfl, football, basketball, tennis, formula-one]
                image:
                  type: string
                  format: binary
                  description: Optional image file for the blog post.
                video:
                  type: string
                  format: binary
                  description: Optional video file for the blog post.
      responses:
        '201':
          description: Blog post created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Blog post created successfully.
                  post:
                    $ref: '#/components/schemas/BlogPost'
        '400':
          description: Bad request. Missing required fields or invalid category.
        '401':
          description: Unauthorized. Admin privileges are required.
        '500':
          description: Internal server error.
  /blogsPost:
    get:
      summary: Fetch all blog posts
      description: Retrieves all blog posts sorted by recency, with optional category filtering.
      tags:
        - Blog Posts
      parameters:
        - name: category
          in: query
          required: false
          description: Filter posts by category (e.g., football, basketball, tennis, etc.). Use "all" to fetch all posts.
          schema:
            type: string
            example: football
      responses:
        '200':
          description: List of blog posts retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlogPost'
        '400':
          description: Invalid category.
        '500':
          description: Internal server error.
  /blogsPost/{category}:
    get:
      summary: Fetch blog posts by specific category
      description: Retrieves all blog posts for a specific category, sorted by recency.
      tags:
        - Blog Posts
      parameters:
        - name: category
          in: path
          required: true
          description: The category of the blog posts to fetch (e.g., football, basketball, etc.).
          schema:
            type: string
            example: football
            enum: [npfl, football, basketball, tennis, formula-one]
      responses:
        '200':
          description: List of blog posts retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BlogPost'
        '400':
          description: Invalid category.
        '500':
          description: Internal server error.
  /blogpost/{postId}:
   delete:
    summary: Delete a blog post
    description: Deletes a blog post by its ID.
    tags:
      - Blog Posts
    parameters:
      - name: postId
        in: path
        description: ID of the blog post to delete
        required: true
components:
  schemas:
...
    BlogPost:
      type: object
      properties:
        _id:
          type: string
          example: "64a1234567bc89def1234567"
        title:
          type: string
          example: "Latest Football Updates"
        description:
          type: string
          example: "This is an update about recent football matches."
        category:
          type: string
          example: football
        image:
          type: string
          example: image1.jpg
        video:
          type: string
          example: video1.mp4
        createdBy:
          type: string
          example: "64a9876543bc21def9876543"
        createdAt:
          type: string
          format: date-time
          example: "2025-01-07T14:30:00.000Z"
...